Before ConEmu

+++ republish.bat
@@ -1,1 +1,1 @@
-npm publish && npm uninstall -g my_wins && npm install -g my_wins && my_wins
+version-select && npm publish && npm uninstall -g my_wins && npm install -g my_wins && my_wins
